Python 3 Compatibility fixes

See merge request mpifr-bdg/mpikat!27

+++ mpikat/core/scpi_client.py
@@ -78,7 +78,7 @@ if __name__ == "__main__":
     usage = "usage: %prog [options]"
     parser = OptionParser(usage = usage)
     parser.add_option('-H', '--host', dest='host', type=str, help='IP to send to')
-    parser.add_option('-p', '--port', dest='port', type=long, help='Port number to send to')
+    parser.add_option('-p', '--port', dest='port', type=int, help='Port number to send to')
     parser.add_option('-m', '--msg', dest='msg',type=str, help='Message to send')
     parser.add_option('-t', '--timeout', dest='timeout',type=float, help='Timeout for message send')
     parser.add_option('', '--log-level',dest='log_level',type=str, help='Set the logging level',default="INFO")
